memblk.cpp MemCells::Read,Writeのアドレス指定をWORDからDWORDに変更 64KB以上アクセス可能に

+++ src/memblk.h
@@ -55,8 +55,8 @@ public:
 	bool SetData( const P6VPATH& );					// ROMデータをファイルから読込み
 	size_t Size() const;							// サイズ(メモリセル数)取得
 	
-	BYTE Read( WORD ) const;						// メモリリード
-	void Write( WORD, BYTE );						// メモリライト
+	BYTE Read( DWORD ) const;						// メモリリード
+	void Write( DWORD, BYTE );						// メモリライト
 	
 	MemCell& operator()( const int );				// メモリセル取得 - operator ()
 };
